requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy>=2.0.30
psycopg[binary]==3.2.12
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6